refactor: consolidate token export
