+++ packages/gill/src/programs/index.ts
@@ -9,14 +9,8 @@ export * from "@solana-program/address-lookup-table";
 export * from "@solana-program/compute-budget";
 export * from "./compute-budget";
 
-export * from "./token-shared";
-
 /**
  * Codama generated clients, stored internally in this package
  * (and associated helpers for them)
  */
 export * from "./token-metadata";
-export * from "./create-token-instructions";
-export * from "./create-token-transaction";
-export * from "./mint-tokens-instructions";
-export * from "./mint-tokens-transaction";
