+++ packages/gill/src/programs/token/addresses.ts
@@ -2,7 +2,7 @@ import type { Address } from "@solana/addresses";
 import type { KeyPairSigner } from "@solana/signers";
 import { findAssociatedTokenPda } from "@solana-program/token-2022";
 import { TOKEN_2022_PROGRAM_ADDRESS } from "@solana-program/token-2022";
-import { checkedAddress } from "../core/utils";
+import { checkedAddress } from "../../core/utils";
 
 export const TOKEN_PROGRAM_ADDRESS =
   "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA" as Address<"TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA">;
