+++ packages/gill/src/programs/token/instructions/create-token.ts
@@ -2,8 +2,11 @@ import type { IInstruction } from "@solana/instructions";
 import type { Address } from "@solana/addresses";
 import type { KeyPairSigner } from "@solana/signers";
 import { getCreateAccountInstruction } from "@solana-program/system";
-import { checkedAddress, getMinimumBalanceForRentExemption } from "../core";
-import { getTokenMetadataAddress, getCreateMetadataAccountV3Instruction } from "./token-metadata";
+import { checkedAddress, getMinimumBalanceForRentExemption } from "../../../core";
+import {
+  getTokenMetadataAddress,
+  getCreateMetadataAccountV3Instruction,
+} from "../../token-metadata";
 
 import {
   extension,
@@ -13,12 +16,12 @@ import {
   getInitializeTokenMetadataInstruction,
   getInitializeMetadataPointerInstruction,
 } from "@solana-program/token-2022";
-import { checkedTokenProgramAddress, type TOKEN_PROGRAM_ADDRESS } from "./token-shared";
+import { checkedTokenProgramAddress, type TOKEN_PROGRAM_ADDRESS } from "../addresses";
 
 export type GetCreateTokenInstructionsArgs = {
   /** Signer that will pay for the rent storage deposit fee */
   payer: KeyPairSigner;
-  /** Token Mint to be created (aka token address) */
+  /** Token mint to be created (aka token address) */
   mint: KeyPairSigner;
   /**
    * The number of decimal places this token should have
@@ -68,7 +71,7 @@ export type GetCreateTokenInstructionsArgs = {
    * ```
    *
    * @example
-   * For `TOKEN_2022_PROGRAM_ADDRESS` use the Mint's address:
+   * For `TOKEN_2022_PROGRAM_ADDRESS` use the mint's address:
    * ```
    * metadataAddress: mint.address;
    * ```
@@ -85,7 +88,7 @@ export type GetCreateTokenInstructionsArgs = {
 };
 
 /**
- * Create the instructions required to initialize a new token's Mint
+ * Create the instructions required to initialize a new token's mint
  */
 export function getCreateTokenInstructions(args: GetCreateTokenInstructionsArgs): IInstruction[] {
   args.tokenProgram = checkedTokenProgramAddress(args.tokenProgram);
