--- packages/gill/src/programs/token22.ts
@@ -1,8 +1,0 @@
-/**
- * Re-export SPL token extension program
- *
- * Note: This is not re-exported from the `./index.ts`
- * file due to a naming collision with original token program
- */
-
-export * from "@solana-program/token-2022";
